requirements.txt
# Core Framework (Stable 0.3 Line)
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0

# Integration Packages (Aligned with 0.3)
langchain-openai>=0.2.0,<0.3.0
langchain-huggingface>=0.0.1,<0.1.0
langchain-ollama>=0.1.0,<0.2.0

# Web & Data
flask==3.1.1 
pypdf==5.6.1 
python-dotenv==1.1.0
chromadb>=0.4.24
sentence-transformers>=3.0.0
httpx>=0.27.0